Restructure project, make requirements explicit
The microraiden python package was moved one level up to the
root directory. The setup.py for microraiden now resides in the
root directory.
Parsing of a requirements-file in setup.py was replaced by an
explicit list of version fixes, as needed for setuptools 'install_requires'.
Only deployment/development specific packages that are not neccessarily
needed for installation should be specified in requirements files.

Also, there where install-scripts added in the Makefile for easy
user installation.